Migrated SpotifyWidget
