+++ package.json
@@ -26,12 +26,12 @@
     "@nanostores/react": "^0.4.0",
     "@react-three/drei": "^9.32.7",
     "@react-three/fiber": "^8.8.8",
+    "@svelte-drama/swr": "^0.2.8",
     "astro-seo": "^0.6.0",
     "comlink": "^4.3.1",
     "cross-fetch": "^3.1.5",
     "nanostores": "^0.7.0",
     "svelte": "^3.52.0",
-    "swr": "^1.3.0",
     "three": "^0.145.0"
   },
   "pnpm": {
